json log exporter for live logs (#7)
This PR introduces logging into file for live logs. The log exporter is attached to the same provider as the other ones.
The file path is given by an env variable which must be present and even the file must be present for the exporter to be registered with the provider.
The file path resides in LIVE_LOGS_FILE_PATH env var, this will be injected into the notebook server.
tested on local, staging, production using notebooks
---------

Co-authored-by: Ankit Saini <asaini@singlestore.com>